util: Fix phase timing error in make_timeline.ts (#4657)

+++ util/logtools/make_timeline.ts
@@ -474,7 +474,7 @@ const assembleTimelineStrings = (
       const ability = phase.split(':')[0];
       const time = phase.split(':')[1];
       if (ability !== undefined && time !== undefined)
-        phases[ability] = parseInt(time);
+        phases[ability] = parseFloat(time);
     }
   }
